+++ core/src/com/game/Manager/User_settings.java
@@ -2,6 +2,6 @@ package com.game.Manager;
 
 public class User_settings {
     public String resolution;
-    public int volume;
+    public float volume;
     public String language;
 }
